requirements.txt
asgiref==3.7.2
Django==4.2.5
django-nested-admin==4.0.2
lxml==4.9.3
Pillow==10.0.1
python-monkey-business==1.0.0
reportlab==4.0.5
six==1.16.0
sqlparse==0.4.4
